chore: rename eks and distribution folders in apis folder to match their provider's names
