+++ internal/apis/kfd/v1alpha2/ekscluster/common/errors.go
