+++ internal/apis/kfd/v1alpha2/ekscluster/common/preflight.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/sighupio/fury-distribution/pkg/apis/ekscluster/v1alpha2/private"
 	"github.com/sighupio/furyctl/configs"
-	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/eks/vpn"
+	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/ekscluster/vpn"
 	"github.com/sighupio/furyctl/internal/cluster"
 	"github.com/sighupio/furyctl/internal/template"
 	"github.com/sighupio/furyctl/internal/tool/awscli"
