+++ internal/apis/kfd/v1alpha2/ekscluster/create/preflight.go
@@ -14,9 +14,9 @@ import (
 
 	"github.com/sighupio/fury-distribution/pkg/apis/config"
 	"github.com/sighupio/fury-distribution/pkg/apis/ekscluster/v1alpha2/private"
-	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/eks/common"
-	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/eks/rules"
-	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/eks/vpn"
+	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/ekscluster/common"
+	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/ekscluster/rules"
+	"github.com/sighupio/furyctl/internal/apis/kfd/v1alpha2/ekscluster/vpn"
 	"github.com/sighupio/furyctl/internal/cluster"
 	"github.com/sighupio/furyctl/internal/diffs"
 	"github.com/sighupio/furyctl/internal/state"
