+++ internal/apis/kfd/v1alpha2/ekscluster/init.go
@@ -2,10 +2,10 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package distribution
+package ekscluster
 
 import (
-	"github.com/sighupio/fury-distribution/pkg/apis/kfddistribution/v1alpha2/public"
+	"github.com/sighupio/fury-distribution/pkg/apis/ekscluster/v1alpha2/private"
 	"github.com/sighupio/furyctl/internal/cluster"
 )
 
@@ -13,13 +13,13 @@ import (
 func init() {
 	cluster.RegisterCreatorFactory(
 		"kfd.sighup.io/v1alpha2",
-		"KFDDistribution",
-		cluster.NewCreatorFactory[*ClusterCreator, public.KfddistributionKfdV1Alpha2](&ClusterCreator{}),
+		"EKSCluster",
+		cluster.NewCreatorFactory[*ClusterCreator, private.EksclusterKfdV1Alpha2](&ClusterCreator{}),
 	)
 
 	cluster.RegisterDeleterFactory(
 		"kfd.sighup.io/v1alpha2",
-		"KFDDistribution",
-		cluster.NewDeleterFactory[*ClusterDeleter, public.KfddistributionKfdV1Alpha2](&ClusterDeleter{}),
+		"EKSCluster",
+		cluster.NewDeleterFactory[*ClusterDeleter, private.EksclusterKfdV1Alpha2](&ClusterDeleter{}),
 	)
 }
