+++ internal/apis/kfd/v1alpha2/ekscluster/rules/extractor.go
