+++ internal/apis/kfd/v1alpha2/ekscluster/test/schema/invalid.yaml
