+++ internal/apis/kfd/v1alpha2/ekscluster/test/schema/min_equal_to_max.yaml
