+++ internal/apis/kfd/v1alpha2/ekscluster/test/schema/min_greater_than_max.yaml
