+++ internal/apis/kfd/v1alpha2/ekscluster/test/schema/min_lesser_than_max.yaml
