+++ internal/apis/kfd/v1alpha2/kfddistribution/delete/preflight.go
