+++ internal/apis/kfd/v1alpha2/kfddistribution/rules/extractor.go
