+++ internal/apis/kfd/v1alpha2/kfddistribution/schema.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package distribution
+package kfddistribution
 
 type ExtraSchemaValidator struct{}
